third_party/blink/web_tests/platform/generic/fast/loader/form-state-restore-with-locked-back-forward-list-expected.txt

PASS location.search is "?TestName1=Ok1&TestName2=Ok2"
PASS successfullyParsed is true

TEST COMPLETE
